refactored Modal calculations & added tabindex, added additional test, added styling to ListCards, tabindex to emoji div
